# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards management system
  description: API Description for Bank Cards management system
  contact:
    name: Alex VM
    email: my@email.com
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/auth/sign-up:
    post:
      tags:
      - Authentication
      summary: Создаётся новый Юзер
      description: |
        Роль нового юзера по умолчанию: 'USER'. Имя и фамилия автоматически используются
        в поле 'CARDHOLDER' для карты, поэтому должны содержать только латинские буквы.
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegisterDto"
        required: true
      responses:
        "201":
          description: User added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRespDto"
        "400":
          $ref: "#/components/responses/Validation"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/auth/login:
    post:
      tags:
      - Authentication
      summary: Login with email and password
      description: Get user ID and JWT token
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLoginDto"
        required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLoginRespDto"
        "400":
          $ref: "#/components/responses/Validation"
...
components:
  schemas:
...
    ErrorResponseSchema:
      type: object
      description: Unified error response
      properties:
        type:
          type: string
          description: "The string containing a URI reference [URI] that identifies\
            \ the problem type."
          example: about:blank
        title:
          type: string
          example: VALIDATION_ERROR
        status:
          type: integer
          format: int32
          description: The number indicating the HTTP status code
          example: 400
        detail:
          type: string
          example: Validation failed
        instance:
          type: string
          description: The string containing a URI reference that identifies the specific
            occurrence of the problem
          example: /users/admin
        errors:
          type: object
          additionalProperties:
            type: string
          description: Validation errors (only for VALIDATION_ERROR)
          example:
            email: must be a valid email
...
    UserRegisterDto:
      type: object
      description: DTO для регистрации нового пользователя
      properties:
        email:
          type: string
          format: email
          description: email
          example: email@email.com
          minLength: 1
        password:
          type: string
          description: password
          example: 12345678
          maxLength: 20
          minLength: 6
        firstName:
          type: string
          description: First Name to print it on card
          example: TOM
          maxLength: 20
          minLength: 2
          pattern: "[A-Za-z]+"
        lastName:
          type: string
          description: First Name to print it on card
          example: SMITH
          maxLength: 20
          minLength: 2
          pattern: "[A-Za-z]+"
      required:
      - email
      - firstName
      - lastName
      - password
    UserRespDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        role:
          type: string
          enum:
          - USER
          - ADMIN
        firstName:
          type: string
        lastName:
          type: string
    UserLoginRespDto:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        token:
          type: string
    UserLoginDto:
      type: object
      description: Аутентификация пользователя
      properties:
        email:
          type: string
          format: email
          description: email
          example: email@email.com
          minLength: 1
        password:
          type: string
          description: password
          example: 12345678
          maxLength: 20
          minLength: 6
      required:
      - email
      - password
...
  responses:
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
    Validation:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponseSchema"
  securitySchemes:
    JWT Bearer:
      type: http
      description: JWT in header
      in: header
      scheme: bearer
      bearerFormat: JWT